requirements.txt
openai
pdfplumber
python-dotenv
reportlab==4.4.4
PyMuPDF
